First round of BarCharts w/ point detail dialog
Includes bar chart features (auto sizing, point detail dialog, axis,
etc.)

testPage.html is just used for testing, it's not necessary for the
library to be functional. On the test page, use "d" to toggle debug
window, "a" to add a random point to the chart, and "r" to remove a
random point from the chart.